# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: GameDash API
  version: 0.1.0
  description: Baseline contract for GameDash MVP foundation.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
paths:
...
  /players/me/profile:
    get:
      operationId: getMyPlayerProfile
      tags: [Players]
      summary: Get current player profile
      responses:
        "200":
          description: Current player profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerProfileResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    patch:
      operationId: updateMyPlayerProfile
      tags: [Players]
      summary: Update current player profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlayerProfileRequest"
      responses:
        "200":
          description: Updated player profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerProfileResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /players/ranks/config:
    get:
      operationId: getRankConfig
      tags: [Players]
      summary: Get MMR to rank mapping
      responses:
        "200":
          description: Rank configuration
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RankConfig"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /players/progression/rewards:
    get:
      operationId: getProgressionRewards
      tags: [Players]
      summary: Get account level reward configuration
      responses:
        "200":
          description: Level reward configuration
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LevelReward"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /players/progression/rules:
    get:
      operationId: getProgressionRules
      tags: [Players]
      summary: Get account progression XP rules
      responses:
        "200":
          description: XP and level threshold configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgressionRulesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /players/{playerId}/progression:
    get:
      operationId: getPlayerProgression
      tags: [Players]
      summary: Get player account progression
      parameters:
        - $ref: "#/components/parameters/playerId"
      responses:
        "200":
          description: Player progression
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerProgressionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /players/{playerId}/mmr:
    get:
      operationId: getPlayerMmr
      tags: [Players]
      summary: Get player MMR by mode
      parameters:
        - $ref: "#/components/parameters/playerId"
      responses:
        "200":
          description: Player MMR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerMmrResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /players/{playerId}/matches:
    get:
      operationId: getPlayerMatches
      tags: [Players]
      summary: Get player match history
      parameters:
        - $ref: "#/components/parameters/playerId"
      responses:
        "200":
          description: Match history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MatchHistoryItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    playerId:
      name: playerId
      in: path
      required: true
      schema:
        type: string
...
  responses:
...
    Unauthorized:
      description: Missing or invalid credentials.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    GameMode:
      type: string
      enum: [ranked, unranked, fun]
...
    PlayerProfileResponse:
      type: object
      required: [userId, pseudo]
      properties:
        userId:
          type: string
        pseudo:
          type: string
        avatarUrl:
          type: string
          format: uri
        region:
          type: string
        bio:
          type: string
    UpdatePlayerProfileRequest:
      type: object
      properties:
        pseudo:
          type: string
        avatarUrl:
          type: string
          format: uri
        region:
          type: string
        bio:
          type: string
...
    PlayerMmrResponse:
      type: object
      required: [playerId, ratings]
      properties:
        playerId:
          type: string
        ratings:
          type: array
          items:
            type: object
            required: [mode, mmr, rank]
            properties:
              mode:
                $ref: "#/components/schemas/GameMode"
              mmr:
                type: integer
              rank:
                type: string
    MatchHistoryItem:
      type: object
      required: [matchId, mode, createdAt]
      properties:
        matchId:
          type: string
          format: uuid
        mode:
          $ref: "#/components/schemas/GameMode"
        createdAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
        result:
          type: string
          enum: [win, loss, draw]
        opponentPlayerId:
          type: string
        mmrBefore:
          type: integer
        mmrAfter:
          type: integer
        mmrDelta:
          type: integer
        rankBefore:
          type: string
        rankAfter:
          type: string
    RankConfig:
      type: object
      required: [mode, minMmr, rank, sortOrder]
      properties:
        mode:
          $ref: "#/components/schemas/GameMode"
        minMmr:
          type: integer
          minimum: 0
        maxMmr:
          type: integer
          minimum: 0
        rank:
          type: string
        sortOrder:
          type: integer
    RewardType:
      type: string
      enum: [soft_currency, cosmetic, title]
    LevelReward:
      type: object
      required: [level, code, label, rewardType]
      properties:
        level:
          type: integer
          minimum: 1
        code:
          type: string
        label:
          type: string
        rewardType:
          $ref: "#/components/schemas/RewardType"
        quantity:
          type: integer
          minimum: 0
    GrantedLevelReward:
      allOf:
        - $ref: "#/components/schemas/LevelReward"
        - type: object
          required: [grantedAt]
          properties:
            grantedAt:
              type: string
              format: date-time
    LevelThreshold:
      type: object
      required: [level, minLifetimeXp]
      properties:
        level:
          type: integer
          minimum: 1
        minLifetimeXp:
          type: integer
          minimum: 0
    ProgressionRulesResponse:
      type: object
      required: [baseXpByMode, outcomeXp, levelThresholds]
      properties:
        baseXpByMode:
          type: object
          required: [ranked, unranked, fun]
          properties:
            ranked:
              type: integer
              minimum: 0
            unranked:
              type: integer
              minimum: 0
            fun:
              type: integer
              minimum: 0
        outcomeXp:
          type: object
          required: [win, loss, draw]
          properties:
            win:
              type: integer
              minimum: 0
            loss:
              type: integer
              minimum: 0
            draw:
              type: integer
              minimum: 0
        levelThresholds:
          type: array
          items:
            $ref: "#/components/schemas/LevelThreshold"
    PlayerProgressionResponse:
      type: object
      required: [playerId, level, lifetimeXp, currentLevelXp, levelProgressPercent, rewards, updatedAt]
      properties:
        playerId:
          type: string
        level:
          type: integer
          minimum: 1
        lifetimeXp:
          type: integer
          minimum: 0
        currentLevelXp:
          type: integer
          minimum: 0
        nextLevelXp:
          type: integer
          minimum: 0
        xpToNextLevel:
          type: integer
          minimum: 0
        levelProgressPercent:
          type: integer
          minimum: 0
          maximum: 100
        rewards:
          type: array
          items:
            $ref: "#/components/schemas/GrantedLevelReward"
        updatedAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, statusCode, timestamp, path, requestId]
          properties:
            code:
              type: string
              example: bad_request
            message:
              type: string
            statusCode:
              type: integer
            timestamp:
              type: string
              format: date-time
            path:
              type: string
              example: /api/v1/auth/login
            requestId:
              type: string